feat: rename PopOver component
