+++ src/components/index.ts
@@ -5,7 +5,7 @@ export { Checkbox } from './Checkbox'
 export { Chip } from './Chip'
 export { DatePicker } from './DatePicker'
 export { Pagination } from './Pagination'
-export { PopOver } from './Popover'
+export { PopOver } from './PopOver'
 export { Radio } from './Radio'
 export { RestrictedSection } from './RestrictedSection'
 export { SearchInput } from './SearchInput'
